split the php image into three images
This should make tests easier to run, because the console server will
able to depend on, and resolve the webserver, which can be necessary
when using Guzzle to test APIs.
Moreover, it separates the nice and clean php-fpm definition from the
quite dirty console definition, that has nodejs, npm, zsh, and will
probably have more development-related packages in the future.
Repetition is avoided with the php image, that is not intended to be run
at all, but to server as a basis for the 2 other images.